Move TailNetHelper to its own file
